renterutil: Support large Writes
Previously Write could only handle buffers up to its MaxChunkSize.
We solve this in the simplest possible fashion: by breaking p into
multiple buffers of at most MaxChunkSize each. This works, but
suffers from an inefficiency: the first such buffer should not be
MaxChunkSize, but rather the maximum number of bytes that could fit
in the current sectors. Anyway, it gets the job done.

Oh, also massively refactor everything; mainly, split all the tricky
(i.e. non-boilerplate) code into its own file, and get rid of a layer
of indirection (all the fd- methods). In Urbit terms, this code is
still quite hot.
